Move utils.group to EquiTrack.utils
