+++ EquiTrack/audit/models.py
@@ -20,6 +20,7 @@ from model_utils.models import TimeStampedModel
 from ordered_model.models import OrderedModel
 
 from EquiTrack.utils import get_environment
+from EquiTrack.utils.wrappers import GroupWrapper
 from attachments.models import Attachment
 from audit.purchase_order.models import AuditorStaffMember, PurchaseOrder, PurchaseOrderItem
 from audit.transitions.conditions import (
@@ -31,7 +32,6 @@ from notification.models import Notification
 from partners.models import PartnerStaffMember, PartnerOrganization
 from utils.common.models.fields import CodedGenericRelation
 from utils.common.urlresolvers import build_frontend_url
-from utils.groups.wrappers import GroupWrapper
 from utils.permissions.models.models import StatusBasePermission
 from utils.permissions.models.query import StatusBasePermissionQueryset
 from utils.permissions.utils import has_action_permission
