+++ EquiTrack/tpm/tests/base.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from utils.groups.wrappers import GroupWrapper
+from EquiTrack.utils.wrappers import GroupWrapper
 
 
 class TPMTestCaseMixin(object):
